型レベルでのswitch-case機構をmisskey-jsからfrontendに持ち込めるようにした
